fix: make sitemap accessible in the public build
